fix tests on windows

+++ pandas/tests/io/excel/test_xlrd.py
@@ -49,7 +49,7 @@ def test_nan_in_xls(datapath):
     # GH 54564
     path = datapath("io", "data", "excel", "test6.xls")
 
-    expected = pd.DataFrame({0: np.r_[0, 2], 1: np.r_[1, np.nan]})
+    expected = pd.DataFrame({0: np.r_[0, 2].astype("int64"), 1: np.r_[1, np.nan]})
 
     result = pd.read_excel(path, header=None)
